backend/requirements.txt
# Recipe Catalog Backend Dependencies

# Core Framework
fastapi==0.116.1
uvicorn[standard]==0.35.0
starlette==0.47.2

# Database
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10

# Authentication & Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
cryptography==45.0.6

# Data Validation
pydantic==2.11.7
pydantic_core==2.33.2
email-validator==2.2.0

# File Upload & Processing
python-multipart==0.0.20
Pillow==10.4.0

# System Monitoring
psutil==7.0.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2

# HTTP & Networking
h11==0.16.0
anyio==4.10.0
sniffio==1.3.1
idna==3.10

# Type Checking & Annotations
typing_extensions==4.14.1
annotated-types==0.7.0

# Dependencies
cffi==1.17.1
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.1
greenlet==3.2.4
pyasn1==0.6.1
pycparser==2.22
rsa==4.9.1
six==1.17.0
typing-inspection==0.4.1
